chore(polish): resolve analyzer warnings, unify model getters, and polish accessibility

+++ lib/widgets/property_card.dart
@@ -413,7 +413,7 @@ class _PropertyCardState extends State<PropertyCard> {
                       tenant: tenant,
                       unitName: unit.name,
                       propertyName: widget.property.name,
-                      rentAmount: tenant.rentAmount,
+                      rentAmount: tenant.rentAmount.toDouble(),
                     );
                   },
                 ),
